fix: 根除unistd.h编译错误 — 移除mtt_internal.h中不必要的<unistd.h>
根因: MTT_DIAG_WRITE宏之前放在mtt_internal.h中，#include <unistd.h>
      导致所有编译单元都需要交叉工具链的unistd.h，而ct-ng工具链无sysroot。

修复:
- mtt_internal.h: 移除 #include <unistd.h>, ssize_t→long
- 每个.c文件自带 #include <unistd.h> (hooks/tracker/reporter/http_server已齐全)
- Makefile恢复原始简洁风格 CROSS_COMPILE ?=
- 本机编译零错误零警告，36/36 PASS

用法(与4天前一模一样):
  export CROSS_COMPILE=/path/to/arm-gcc13-linux-gnueabi-
  export ARCH=arm
  make

+++ src/mtt_internal.h
@@ -12,13 +12,14 @@
 #define MTT_INTERNAL_H
 
 #include <stddef.h>
-#include <unistd.h>
+/* 禁用unistd.h: 交叉编译器自带路径, 不引入额外头文件依赖 */
 
 /* 安全忽略 write() 返回值（诊断输出场景，失败无影响）。
  * GCC 14+ / 13+ 中 (void)write() 无法抑制 warn_unused_result，
- * 必须通过实际捕获返回值来消除警告。 */
+ * 必须通过实际捕获返回值来消除警告。
+ * 调用者需自行 #include <unistd.h> 获取 write() 声明。 */
 #define MTT_DIAG_WRITE(fd, buf, len) \
-    do { ssize_t __mtt_w = write((fd), (buf), (len)); (void)__mtt_w; } while(0)
+    do { long __mtt_w = (long)write((fd), (buf), (len)); (void)__mtt_w; } while(0)
 #include <stdint.h>
 #include <time.h>
 #include <pthread.h>
